Refactor Django tutorial documentation
The Django tutorial in the docs has been simplified and made more general. The specific 'music' module has been removed from the instructions, replacing it with a guideline to check whether an app is included in the 'INSTALLED_APPS' setting. In addition, changes to the title format and other minor text changes have been implemented to improve clarity and readability.